refactor(ui): split api.tsx into domain modules
Move dataset, feed/bird food, cameras/health, file-test, settings session, audit/metrics, push/processor weights, yaml/db, species registry hub, and species/overview/detections out of api.tsx; keep public API via re-exports. Refs #343

Made-with: Cursor
