+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
-- **[#343](https://github.com/Gfermoto/BirdLense-Hub/issues/343) (фаза A):** UI — **`client.ts`** (base URL, axios timeout, **`JOB_STATUS_POLL_TIMEOUT_MS`**), **`timeline.ts`**, **`video.ts`** (видео, соседи, fusion trace, nearest recording day, delete/regen/merge), **`weatherRegion.ts`**, **`migrationCalendar.ts`**; реэкспорты из **`api.tsx`**; **`queryKeys`** — таймлайн, unknowns, **video**, bird-directory, **speciesSummary** (в т.ч. `bySpecies`), **storage** (stats), **overview** / **weather** / **calendar**, **`systemPanels`** (аудит конфига, observability, каталог, recognition/fusion, веса и др.), **feed**, **health** (статус шапки), **live** (камеры), **birdFood**, **fileTest**, **speciesDirectory**; миграция ключей на **Timeline**, **Unknowns**, **VideoDetails** / **VideoInfo** / **DetectedSpecies**, **Overview**, **WeatherCard**, **MigrationCalendar**, **DatabaseMaintenanceCard**, **StorageOverview**, **RecordingsCalendar**, **DatasetExportsCard**, **Navigation**, **FeedCard**, **StatusIndicator**, **Footer**, **Live**, **FoodManagement**, **FileReplayCard**, карточки **Система** (**SystemHero**, **ConfigAudit**, **Observability**, **CatalogRepair**, **RecognitionImprovement**, **SpeciesDataQuality**, **ClassifierDatasetAlignment**, **ProcessorWeights**, **AutomationFusion**), **SpeciesDirectory**, **SpeciesSummary**, **ProcessorFrigateFusionBlock**; тесты **`timeline.test.ts`**, **`video.test.ts`**, **`weatherRegion.test.ts`**.
+- **[#343](https://github.com/Gfermoto/BirdLense-Hub/issues/343) (фаза A):** UI — **`client.ts`** (base URL, axios timeout, **`JOB_STATUS_POLL_TIMEOUT_MS`**), **`timeline.ts`**, **`video.ts`** (видео, соседи, fusion trace, nearest recording day, delete/regen/merge), **`weatherRegion.ts`**, **`migrationCalendar.ts`**, **`dataset.ts`**, **`birdFoodFeed.ts`**, **`camerasHealth.ts`**, **`fileTest.ts`**, **`settingsSession.ts`**, **`systemAuditMetrics.ts`**, **`notificationsProcessor.ts`** (push / notify / restart / веса), **`settingsYamlDb.ts`** (YAML, БД, purge, ZIP→координаты), **`speciesRegistryHub.ts`** (реестр, fusion/recognition, BirdNET FIFO, purge-диагностика), **`speciesOverviewDetections.ts`** (каталог/обзор вида, тюнинг, Xeno-Canto, PDF, детекции, review-queue, corrections); **`api.tsx`** — `getApiErrorMessage`, `resolveImageUrl` и реэкспорты; **`queryKeys`** — таймлайн, unknowns, **video**, bird-directory, **speciesSummary** (в т.ч. `bySpecies`), **storage** (stats), **overview** / **weather** / **calendar**, **`systemPanels`** (аудит конфига, observability, каталог, recognition/fusion, веса и др.), **feed**, **health** (статус шапки), **live** (камеры), **birdFood**, **fileTest**, **speciesDirectory**; миграция ключей на **Timeline**, **Unknowns**, **VideoDetails** / **VideoInfo** / **DetectedSpecies**, **Overview**, **WeatherCard**, **MigrationCalendar**, **DatabaseMaintenanceCard**, **StorageOverview**, **RecordingsCalendar**, **DatasetExportsCard**, **Navigation**, **FeedCard**, **StatusIndicator**, **Footer**, **Live**, **FoodManagement**, **FileReplayCard**, карточки **Система** (**SystemHero**, **ConfigAudit**, **Observability**, **CatalogRepair**, **RecognitionImprovement**, **SpeciesDataQuality**, **ClassifierDatasetAlignment**, **ProcessorWeights**, **AutomationFusion**), **SpeciesDirectory**, **SpeciesSummary**, **ProcessorFrigateFusionBlock**; тесты **`timeline.test.ts`**, **`video.test.ts`**, **`weatherRegion.test.ts`**.
 
 - **Документация (CONFIGURATION):** EN/RU — блок **On this page** / **По странице** (быстрые якоря), ссылки на **`app/.env.example`** и **`docs/project/openapi.md`** в шапке; **I18N_STATUS** — строка **SETTINGS_TRIGGERS_PHASE2**, шаг про OpenAPI/contract-тест при смене HTTP-маршрутов.
